fix: Header link depth 줄임
